- Internal readers dont support tcflush()
After some testing by Stefansat and me its obvious internal readers dont support flushing -> disable flush capability for internal readers!
The reason it was enabled is that the ATR code didnt read all of the chars to the end leaving a 0x90 0x00 (and another 0x00 on some cards) in the io buffer.
If these chars wont be read or flushed card cant be used.

Now we added some additional reading (9 bytes should be enough I think) after fetching the ATR to get rid of these chars.

@VU boxes manufacturer: please fix your drivers in such a way that if we read with a certain timeout value your reader driver respects it!
If we ask with a timeout of 75ms dont ignore it and make it 4 seconds.
This behaviour observed on vu duo box by stefansat, possibly more boxes are affected!


git-svn-id: http://www.streamboard.tv/svn/oscam/trunk@10067 4b0bc96b-bc66-0410-9d44-ebda105a78c1