Allocate TUNTAB dynamically
The TUNTAB structure size is 602 bytes and also the number of
entries is limited.

This structure is used to implement the beta tunnel feature via
'betatunnel' account setting.

This feature is not used by everyone but everyone takes the memory hit.

This patch makes the bulk of the structure dynamically allocated. The
memory savings are as follows (x86_64 defconfig):

{{{
 struct         before   after   delta  savings
  s_tuntab         602      16    -586    97.4%
  s_client        2440    1856    -584    24.0%
  s_auth          1976    1392    -584    29.6%

 Binary size:
  add/remove: 1/0 grow/shrink: 5/5 up/down: 335/-485 (-150)
}}}

git-svn-id: http://www.streamboard.tv/svn/oscam/trunk@10454 4b0bc96b-bc66-0410-9d44-ebda105a78c1